rename switch_thread -> switch_to_thread & add tests for reference results

+++ async_coro/include/async_coro/internal/store_type.h
@@ -9,6 +9,45 @@
 #endif
 
 namespace async_coro::internal {
+
+template <typename T>
+struct result_coro_type {
+  T result;
+
+  T& get_ref() noexcept { return result; }
+  const T& get_cref() const noexcept { return result; }
+  T move() noexcept {
+    return std::move(result);
+  }
+
+  template <typename... TArgs>
+  void inplace_init(TArgs&&... args) noexcept {
+    new (&this->result) T(std::forward<TArgs>(args)...);
+  }
+
+  void destroy() noexcept(noexcept(std::is_nothrow_destructible_v<T>)) {
+    std::destroy_at(&result);
+  }
+};
+
+template <typename T>
+struct result_coro_type<T&> {
+  T* result;
+
+  T& get_ref() noexcept { return *result; }
+  const T& get_cref() const noexcept { return *result; }
+  T& move() noexcept {
+    // actually we cant move refs here as the dont hold value
+    return *result;
+  }
+
+  void inplace_init(T& ref) noexcept {
+    this->result = &ref;
+  }
+
+  void destroy() noexcept {}
+};
+
 #if ASYNC_CORO_NO_EXCEPTIONS
 template <typename T>
 union store_type {
@@ -21,7 +60,9 @@ union store_type {
   ~store_type() noexcept {}
   void destroy_exception() {}
   void destroy_result() noexcept(noexcept(std::is_nothrow_destructible_v<T>)) {
-    std::destroy_at(&result);
+    if constexpr (!std::is_reference_v<T>) {
+      std::destroy_at(&result);
+    }
   }
 };
 
@@ -42,16 +83,15 @@ union store_type {
       std::is_nothrow_destructible_v<T>;
 
   std::exception_ptr exception;
-  T result;
+  result_coro_type<T> result;
 
   store_type() noexcept {}
   ~store_type() noexcept {}
-  void destroy_exception() noexcept(
-      std::is_nothrow_destructible_v<std::exception_ptr>) {
+  void destroy_exception() noexcept(std::is_nothrow_destructible_v<std::exception_ptr>) {
     std::destroy_at(&exception);
   }
-  void destroy_result() noexcept(noexcept(std::is_nothrow_destructible_v<T>)) {
-    std::destroy_at(&result);
+  void destroy_result() noexcept(noexcept(std::is_reference_v<T> || std::is_nothrow_destructible_v<T>)) {
+    result.destroy();
   }
 };
